pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "adafruit-circuitpython-fingerprint>=2.2.19",
    "fastapi>=0.115.12",
    "firebase-admin>=6.7.0",
    "pydantic>=2.11.1",
    "pyserial>=3.5",
    "python-multipart>=0.0.20",
    "pytz>=2025.2",
    "uvicorn>=0.34.0",
]
